Test1 исправление логов float64 ->numeric(14,2)

+++ internal/server/middleware_DBconn.go
@@ -59,8 +59,8 @@ func CreateTables(FlagDatabaseURI string) {
 			id_order 			text not null unique,
 			processed_at 		timestamptz not null,
     		total_balance_score	numeric(14,2),
-			order_balance_score float64,
-			total_write_off 	float64
+			order_balance_score numeric(14,2),
+			total_write_off 	numeric(14,2)
 			)`)
 	if err != nil {
 		log.Println("Failed to create balance table:", err)
